feat(core): add CalcResult type system, memory/variables, percent, gradian
Phase 1 of Casio fx-991MS parity:
- CalcResult sealed class (Real, Complex, Integer, Matrix, Vector)
- Evaluator refactored from Double to CalcResult return type
- MemoryManager: Ans, 9 variables (A-F,M,X,Y), independent M (M+/M-)
- Percent as postfix operator (x% = x/100)
- Random number generator (Ran#)
- Gradian angle unit (DEG/RAD/GRA)
- Variable/Ans tokens in Lexer, Parser, Evaluator
- All 145 existing tests pass

+++ app/src/main/java/com/calcmate/scientificcalculator/feature/graphing/GraphViewModel.kt
@@ -135,7 +135,7 @@ class GraphViewModel : ViewModel() {
             val substituted = substituteX(expression, x)
             val tokens = Lexer(substituted).tokenize()
             val ast = Parser(tokens).parse()
-            Evaluator().evaluate(ast)
+            Evaluator().evaluate(ast).toDouble()
         } catch (_: Exception) {
             Double.NaN
         }
